Add server_data scripts
This module contains some modifications to dbLoadPlayer.sqf. The spawn point is hard coded, the starting inventory is modified, and there is a critical fix for player hive saving. This module is redirected to the `dzlegacy\server_data` prefix. It appears the "scripts" folder has been removed. This version compared to vanilla has scripts for cooking, as well as spawning heli crashes and wild animals, although these are not working yet. It appears that, unintentionally, I have duplicated `player_queued.sqf` from `modules_Dayz` and added it to server_data_legacy instead. This will need to be remedied accordingly for when the DayZ Legacy Mission makes the transition to using exclusively redirected scripts from `dzlegacy\modules_dayz` rather than the vanilla addon.
The init.sqf for this addon is located and called within the server mission at initServer.sqf.