Fix player position kickbacks
This issue was caused by my script redirections. I inadvertently created a variable scope issue which kicked my player out of any vehicles when it attempts to poll information related to the variable it no longer has a scope on. This fix is being combined with some changes to the DayZ Legacy Mission and how it calls redirected scripts.